% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fxns.R
\name{add_month_name}
\alias{add_month_name}
\title{Add month name to a data frame}
\usage{
add_month_name(df)
}
\value{
a data frame that contains a column of the month name
}
\description{
\code{add_month_name} Adds the month name to a data frame that contains month
number
}
