% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo.concatenate}
\alias{cdo.concatenate}
\title{Concatenate netcdfs and convert to absolute time}
\usage{
cdo.concatenate(df, cdo_path, ofile)
}
\arguments{
\item{df}{a data frame containing the a path to the netcdf files to concatenate together}

\item{cdo_path}{the path to the cdo to use, the default is set to the cdo installed on pic}

\item{ofile}{the file to write the concatenated netcdf to, default is set to my scratch file on pic}
}
\value{
a concatenated netcdf with absolute time
}
\description{
\code{cdo.concatenate}
}
\keyword{internal}
