% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo.processor}
\alias{cdo.processor}
\title{Concatenates files CMIP5 netcdfs together and then carries out a single cdo command line}
\usage{
cdo.processor(df, cdo_path = "/share/apps/netcdf/4.3.2/gcc/4.4.7/bin/cdo",
  cdo_command, ofile = "/pic/scratch/dorh012/tfile1")
}
\arguments{
\item{df}{is the dataframe of the file path and cmip5 information}

\item{cdo_path}{is the path of the location of the cdo used to process the netcdfs}

\item{cdo_command}{the name of the cdo command(s) that will be used to process the netcdf}

\item{ofile}{is the path to the location where to save a temporary netcdf}
}
\value{
a data frame of the processed netcdf data and CMIP5 file information
}
\description{
\code{cdo.processor}
}
