% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{cdo.sellonlat}
\alias{cdo.sellonlat}
\title{The R functional form of the cdo operator sellonlat}
\usage{
cdo.sellonlat(area_input, data_input, defined_basins,
  cdo_operator = "fldmean", intermediate_output = BASE_NAME,
  cdo_path = "/share/apps/netcdf/4.3.2/gcc/4.4.7/bin/cdo",
  temp_dir = "/pic/scratch/dorh012/")
}
\arguments{
\item{area_input}{a data frame containing cmip info and the path to the cell area netcdfs}

\item{data_input}{a data frame containing cmip info and the path to the data netcdfs}

\item{defined_basins}{a data frame containing the lat/lon boundaries of an ocean basin}

\item{intermediate_output}{default set to BASE_NAME, it is the defined location where to save any intermediate output}

\item{cdo_path}{the path to the cdo to use}

\item{temp_dir}{the path to the location where to save the intermediate netcdfs from the cdo}
}
\value{
a data frame of the all the processed netcdf data in all of the regions
}
\description{
\code{cdo.sellonlat} is function that applies a cdo operator over a defined
geographical regions.
}
\keyword{internal}
