% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fxns.R
\name{check.column}
\alias{check.column}
\title{Check input dataframes for the required columns}
\usage{
check.column(df, df_name, required_columns)
}
\arguments{
\item{df}{the data frame to check for required columns}

\item{df_name}{the name of the df to return in the error message}

\item{required_columns}{the names of the columns to check for in the data frame}
}
\value{
an intentional error message if the data frame does not pass the tests
}
\description{
\code{check.column} Checks to make sure that any data frames used as an input
contains all of the required columns. If the data frame does not then the a
stop error will occur
}
\keyword{internal}
