% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip5_fxns.R
\name{cmip.check_depth}
\alias{cmip.check_depth}
\title{Check to see if the CMIP5 netcdf file has information about the depth or not.}
\usage{
cmip.check_depth(ncdf_paths)
}
\arguments{
\item{ncdf_paths}{the path to a single or vector of paths to cmip5 files}
}
\value{
a data frame containing the netcdf file path and the Y/N indicator of depth existence
}
\description{
\code{cmip.check_depth} imports a singular netcdf file and then checks to see if
the file contains any information about ocean depth or "lev".
}
