% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip5_fxns.R
\name{cmip.file_info}
\alias{cmip.file_info}
\title{Parse out cmip file information from the cmip netcdf file name}
\usage{
cmip.file_info(file_path)
}
\arguments{
\item{file_path}{a vector containing the path to cmip5 files to process}
}
\value{
a data frame file name and relevant cmip5 file information
}
\description{
\code{cmip.file_info} cmip5 netcdfs should all have the same file name, this
function splits the file name into relevant cmip5 file information such as
model name, ensemble member, experiment, ect.
}
\keyword{internal}
