% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmip5_fxns.R
\name{cmip.find_me}
\alias{cmip.find_me}
\title{Find CMIP5 netcdfs of interest}
\usage{
cmip.find_me(path, variable = "[a-zA-Z0-9-]+", domain = "[a-zA-Z0-9-]+",
  model = "[a-zA-Z0-9-]+", experiment = "[a-zA-Z0-9-]+",
  ensemble = "[a-zA-Z0-9-]+", time = "[0-9]+)-([0-9]+")
}
\arguments{
\item{path}{the directory to search for the CMIP5 files in}

\item{variable}{default is set to search for all letter/number combinations, may specify as a single or multiple search requirement(s)}

\item{domain}{default is set to search for all letter/number combinations, may specify as a single or multiple search requirement(s)}

\item{model}{default is set to search for all letter/number combinations, may specify as a single or multiple search requirement(s)}

\item{experiment}{default is set to search for all letter/number combinations, may specify as a single or multiple search requirement(s)}

\item{ensemble}{default is set to search for all letter/number combinations, may specify as a single or multiple search requirement(s)}

\item{time}{default is set to search for numeric dates, may specify as a single or multiple search requirement(s)}
}
\value{
a list of paths for CMIP5 netcdf files
}
\description{
\code{cmip.find_me} returns a list of the pathes to the CMIP5 netcdf files of interest.
Unless otherwise specified this function is setup to search for any variable, model,
ensemble, domain, date .ncdf files. This function can search for multiple CMIP5 specificaitons
at a time for example if  you are interested in fidning all of the CMIP5 files for the rcp45
and 1pctCO2 experiements the experiment parameter would be specified as experiment = c("rcp45", "1pctCO2").
}
