% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{internal.sellonlat}
\alias{internal.sellonlat}
\title{Internal function that applies a cdo operator to all of the defined geographical regions}
\usage{
internal.sellonlat(internal_df)
}
\arguments{
\item{internal_df}{a data frame created in the cdo.sellonlat function}
}
\value{
a data frame of the processed netcdf data for a single variable/model in all of the regions
}
\description{
\code{internal.sellonlat} is function that is used internally in the cdo.sellonlat function
}
\keyword{internal}
