% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo_fxns.R
\name{internal.sellonlat_cdo}
\alias{internal.sellonlat_cdo}
\title{Applies a cdo operator to a defined geographical region}
\usage{
internal.sellonlat_cdo(basin_df)
}
\arguments{
\item{basin_df}{a data frame of the geographic boundaries of the defined region}
}
\value{
a data frame of the processed netcdf data and CMIP5 file information for a single region
}
\description{
\code{internal.sellonlat_cdo} is function that is used internally in the cdo.sellonlat
function.
}
\keyword{internal}
