% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fxns.R
\name{list.add_list_name}
\alias{list.add_list_name}
\title{Add names to a list stored in a list}
\usage{
list.add_list_name(input_list)
}
\arguments{
\item{input_list}{an object that contains multiple lists that are un-named, requires a column of names}
}
\value{
A list of named figures
}
\description{
\code{list.add_list_name} This function adds names to an object that is stored in a list
}
