% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_fxns.R
\name{vis.add_fig_name}
\alias{vis.add_fig_name}
\title{Add names to figures stored in a list}
\usage{
vis.add_fig_name(fig_list)
}
\arguments{
\item{fig_list}{an object that contains multiple lists that are un-named, requires a column of names}
}
\value{
A list of named figures
}
\description{
\code{vis.add_fig_name} This function adds names to figures that are stored in a list
}
