% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_fxns.R
\name{vis.compare_id}
\alias{vis.compare_id}
\title{Compare differnt data frames by the "id" variable}
\usage{
vis.compare_id(df)
}
\arguments{
\item{df}{a data frame containing the information to plot}
}
\value{
A list containing a figure that compares the two data sets based on the id paramter
}
\description{
\code{vis.compare_id} This function generates line plots based on some variable termed id
}
