% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_fxns.R
\name{vis.exploratory}
\alias{vis.exploratory}
\title{Make exploratory graphs}
\usage{
vis.exploratory(input_df)
}
\arguments{
\item{input_df}{is the data frame of processed cmip5 data to plot}
}
\value{
a list containing boxplots and time series of mean monthly data, includes regular
and normalized figures.
}
\description{
\code{vis.exploratory} Uses an internal graphing function to create figures for
each unique variable/ensemble/experiment combination, I think it can easily be
modified so do comparison groups
}
