% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_fxns.R
\name{vis.is_outlier}
\alias{vis.is_outlier}
\title{Identify outliers}
\usage{
vis.is_outlier(x)
}
\arguments{
\item{x}{is the vector to check for outliers, values that lay outside of the 1.5*IQR}
}
\value{
an indicator varibale x if the the value is an outlier
}
\description{
\code{vis.is_outlier}
}
\keyword{internal}
