% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_fxns.R
\name{vis.time_series}
\alias{vis.time_series}
\title{Make a figure of a concatenated time series}
\usage{
vis.time_series(input_df)
}
\arguments{
\item{df}{aa data frame with information to plot}
}
\value{
A figure of a timeseries
}
\description{
\code{vis.time_series} This function adds names to figures that are stored in a list
}
